NEW Tools/The sims 2/The Sims 2 TXFL Extractor/requirements.txt
ReverseBox
pillow